backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
tensorflow>=2.13.0,<2.17
pillow>=10.0.0
numpy>=1.24.0,<2.0
python-multipart>=0.0.9
scikit-learn>=1.3.0
kagglehub>=0.2.0
kaggle>=1.6.17
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
aiofiles>=23.0.0
tensorflow-metal>=1.1.0; platform_system=="Darwin"
